changes to blocks and page_about
Created new Login and regristration pages. Added the information for the
terms and service page. About us page now has content on it. How it
works is still in progress and how it works for employers has not been
started, however will follow the same formatting of the how it works for
students